Extract interface from URLEngineDownloader
extract common interface for engine downloaders and extract url lookup
from downloader (seperate class, EngineDownloadURLResolver) for better
maintainability
